Add quickfix for nullability annotations with @TypeQualifierDefault
First quickfix is to add nullability annotations which are annotated with @javax.annotation.meta.TypeQualifierDefault to the package-info.java of current package. When there is no package-info.java available it will create one on the fly.

The quickfix takes into account the current element. If the element is not one of the target types of the nullability annotation then fix will not be made available. For example @ParametersAreNonnullByDefault will not be made available as a quickfix for a field.

Also, only those nullability annotations registered with the annotation manager of IntelliJ will be considered.